Added Dashboard_Redesign:
Fixed a bug where the slider would go places, not adapting to the slider's height regarding the resolution of the screen.

I started with the main container. The strategy here will imply both bootstrap and CSS. Class names will be bootstrap keywords, such as "container-fluid". As such, what has keywords as class="" will also have id's under main-box. This will be the way that div's will be referred in the .css file when changing margin, padding and so on.