Updated the front-end to support downloading of files.
Form allows now allows for multiple inputs, which maps it to a special restful service in the ClientResource. It makes a distinction between csv and xlsx, which both can be downloaded.

TODO: We still need to work on the correct downloading, so that it can access the files either way.